Initial commit
This creates a small python script that wraps calls to udica and the
CRI in order to generate policies for specific pods in the kubernetes
cluster.

This comes with some helper scripts and the manifest necessary to deploy
the pod. Currently you have to exec into the pod to run the commands.